feat(tts): add catalog metadata handling for TTS model detection and update CMake configuration
